+++ extension/src/SymbolReference/SymbolReferenceCache.ts
@@ -1,0 +1,29 @@
+import { AppPackage } from './types/AppPackage';
+
+let cachedAppPackages: AppPackage[] = [];
+
+export function addAppPackageToCache(appPackage: AppPackage): void {
+    if (appPackageInCache(appPackage)) {
+        // app already in cache
+        return;
+    }
+
+    let appToCache = appPackage;
+    appToCache.symbolReference = undefined; // Free up unnecessary memory allocation
+    cachedAppPackages.push(appToCache);
+    cachedAppPackages.sort((a, b) => { return a.sort(b); });
+}
+
+export function getAppPackageFromCache(name: string, publisher: string, version: string): AppPackage {
+    let cachedAppPackage = cachedAppPackages.filter(p => p.name === name && p.publisher === publisher && p.version === version)[0];
+    return cachedAppPackage;
+}
+
+function appPackageInCache(appPackage: AppPackage): boolean {
+    return cachedAppPackages.filter(p => p.packageId === appPackage.packageId).length > 0;
+}
+
+export function appInCache(name: string, publisher: string, version: string): boolean {
+    return cachedAppPackages.filter(p => p.name === name && p.publisher === publisher && p.version === version).length > 0;
+}
+